+++ daemon/logger/templates/templates_test.go
@@ -1,0 +1,18 @@
+package templates
+
+import (
+	"bytes"
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+)
+
+func TestNewParse(t *testing.T) {
+	tm, err := NewParse("foo", "this is a {{ . }}")
+	assert.NoError(t, err)
+
+	var b bytes.Buffer
+	assert.NoError(t, tm.Execute(&b, "string"))
+	want := "this is a string"
+	assert.Equal(t, want, b.String())
+}
