+++ profiles/apparmor/apparmor.go
@@ -9,9 +9,9 @@ import (
 	"os"
 	"path"
 	"strings"
+	"text/template"
 
 	"github.com/docker/docker/pkg/aaparser"
-	"github.com/docker/docker/pkg/templates"
 )
 
 var (
@@ -33,7 +33,7 @@ type profileData struct {
 
 // generateDefault creates an apparmor profile from ProfileData.
 func (p *profileData) generateDefault(out io.Writer) error {
-	compiled, err := templates.NewParse("apparmor_profile", baseTemplate)
+	compiled, err := template.New("apparmor_profile").Parse(baseTemplate)
 	if err != nil {
 		return err
 	}
